StatsdClient can now be given a postfix to append to all logging messages, also fixed unit tests
The unit tests were completely broken due to the change to async
everywhere so they are now fixed and Moq is mostly eliminated in favor
of a simple stub object.

StatsdClient can now be given a postfix name which will be appended to
every stat.  If the Statsd Client object is given the prefix
"some.prefix" and the postfix "hostname" then using
client.LogGauge("mygauge", 1) will result in the gauge
"some.prefix.myguage.hostname" being sent to statsd.  The use case for
this feature is based on how Krux does monitoring, see presentation
here:
https://puppetlabs.com/presentations/puppet-camp-london-2014-how-measure-everything-million-metrics-second-minimal